Fix: recreating existing users, Add creator column
Fixed an issue where users were being created in the users database each time they logged in, by modifying the logic in some functions in the signInWithGoogle page (the value of true wasn't being returned from the promise)
Also changed user creation to make the document name be the same as the user's ID
Added another column in datatable to show the Creator of the match, with their display name. Makes it easier to delete scores from a specific user.